requirements.txt
appnope==0.1.0
decorator==4.0.9
Django==1.7.11
django-cors-headers==1.1.0
django-extensions==1.6.1
djangorestframework==3.3.3
gnureadline==6.3.3
ipython==4.1.2
ipython-genutils==0.1.0
model-mommy==1.2.6
nodeenv==0.13.6
pexpect==4.0.1
pickleshare==0.7.2
ptyprocess==0.5.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
Werkzeug==0.11.5
